+++ Scripts/Entities/Run.cs
@@ -22,7 +22,7 @@ namespace SailwindRegatta
         public int? BoatTypeId { get; set; }
 
         // Accumulated real-world play time in seconds. Excludes paused and closed-game time.
-        public float ElapsedSeconds { get; set; }
+        public double ElapsedSeconds { get; set; }
 
         public CheckpointName NextCheckpointName => Race.RouteCheckpoints[NextCheckpointIndex];
         public bool IsFinished => NextCheckpointIndex >= Race.RouteCheckpoints.Length;
